Rename customization.splitname() to utils.parsename().
Function belongs in the utils module, and the renaming both better
describes what the function does, and avoids confusion with the
utils.splitnames() function.
